pyproject.toml
[tool.poetry]
name = "makejinja"
version = "1.0.0b3"
description = "Automatically generate files based on Jinja templates. Use it to easily generate complex Home Assistant dashboards!"
authors = ["Mirko Lenz <mirko@mirkolenz.com>"]
license = "MIT"
readme = "README.md"
repository = "https://github.com/mirkolenz/makejinja"

[tool.poetry.scripts]
makejinja = "makejinja.cli:main"

[tool.poetry.dependencies]
python = "^3.9"
jinja2 = "^3.1.2"
rich = "^13.0.0"
pyyaml = "^6.0"
rich-click = "^1.6.0"
typed-settings = { extras = ["click"], version = "^2.0.0" }
tomli = {version = "^2.0.1", python = "<3.11"}

[tool.poetry.group.dev.dependencies]
pytest = "^7.2.0"
pytest-cov = "^4.0.0"

[tool.pytest.ini_options]
addopts = "--cov makejinja --cov-report term-missing -vv"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
